test_requirements.txt
BeautifulSoup==3.2.1
Django==1.5.5
PyJWT==0.2.1
South==0.8.1
coverage==3.6
django-admin-tools==0.5.1
django-appconf==0.6
django-apptemplates==0.0.1
django-compressor==1.2
django-filetransfers==0.0.0
django-fluent-contents==0.9a1
django-localflavor==1.0
django-loginas==0.1.3
django-nose==1.2
django-polymorphic==0.5.6
django-registration==1.0
django-setuptest==0.1.4
django-social-auth==0.7.23
django-tag-parser==1.1.0
django-taggit==0.10a1
django-template-analyzer==1.3
django-templatetag-handlebars==1.2.0
django-tinymce==1.5.1b2
django-tools==0.25.0
django-wysiwyg==0.5.1
djangorestframework==2.3.12
# djangorestframework-jwt==0.1.5
factory-boy==2.3.1
httplib2==0.9
logilab-astng==0.24.3
logilab-common==0.62.1
micawber==0.2.6
mock==1.0.1
mollie-api-python==1.0.0
nose==1.3.3
oauth2==1.5.211
pep8==1.5.7
pylint==1.1.0
python-openid==2.2.5
requests==2.3.0
sauceclient==0.1.0
selenium==2.42.1
six==1.7.3
sorl-thumbnail==11.12
splinter==0.6.0
wsgiref==0.1.2

# JSON Web Token based authentication for Django REST framework
-e git+https://github.com/onepercentclub/django-rest-framework-jwt@master#egg=djangorestframework-jwt-0.1.5
